pyproject.toml
[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[project]
name = "nemorosa"
version = "0.0.1"
description = "A specialized cross-seeding tool designed for music torrents, featuring automatic file mapping, partial matching, and seamless torrent injection"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "KyokoMiki" },
]
keywords = ["torrent", "gazelle", "cross-seeding"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "colorama>=0.4.6",
    "deluge-client>=1.10.2",
    "humanfriendly>=10.0",
    "msgspec>=0.18.6",
    "platformdirs>=4.4.0",
    "pyyaml>=6.0.2",
    "qbittorrent-api>=2025.7.0",
    "requests>=2.32.5",
    "torf>=4.3.0",
    "transmission-rpc>=7.0.11",
]

[project.urls]
Homepage = "https://github.com/KyokoMiki/nemorosa"
Repository = "https://github.com/KyokoMiki/nemorosa"
Issues = "https://github.com/KyokoMiki/nemorosa/issues"

[project.scripts]
nemorosa = "nemorosa:main"

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
